Code refactorization.
TimeCardsManager model has been finalized without listener, It has a dynamic
list and also updates the database concurrently. Another big change is that the
AddCardDialogFragment and GraphDetailFragment are child Fragments of
TimeCardsFragment. MainActivity orchestrates between SettingsFragment and
TimeCardsFragment.

Lastly, TimeCard no longer has points and count as members rather there is a
Cache class to imply that these members are related to the TimeCard but should
not be relied upon since they can be null, -1, etc. So now to get this
information populated and up to date, make sure TimeCardsManager.query has been
called with the desired TimeCard.